# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Research Agent API",
    "description": "Production-grade REST API for the Research Agent platform. Supports academic literature search, document ingestion, retrieval-augmented generation (RAG) question answering, summarization, and similarity analysis.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/ingest": {
      "post": {
        "tags": [
          "ingest"
        ],
        "summary": "Search, download, process, and index articles",
        "description": "Runs the full ingest pipeline: search \u2192 download \u2192 process \u2192 index.",
        "operationId": "ingest_articles",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ingest/download": {
      "post": {
        "tags": [
          "ingest"
        ],
        "summary": "Search and download articles without indexing",
        "description": "Search for articles and download PDFs without processing or indexing.",
        "operationId": "search_and_download",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DownloadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/documents": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "List indexed documents",
        "description": "Returns paginated list of indexed documents with metadata.",
        "operationId": "list_documents",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filter by title",
              "default": "",
              "title": "Query"
            },
            "description": "Filter by title"
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Pagination cursor",
              "title": "Cursor"
            },
            "description": "Pagination cursor"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page",
              "default": 20,
              "title": "Limit"
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Get document details",
        "description": "Returns document metadata and chunks by ID.",
        "operationId": "get_document",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Document ID",
              "title": "Document Id"
            },
            "description": "Document ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}/chunks": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Get document chunks",
        "description": "Returns paginated chunks for a document.",
        "operationId": "get_document_chunks",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Document ID",
              "title": "Document Id"
            },
            "description": "Document ID"
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Pagination cursor",
              "title": "Cursor"
            },
            "description": "Pagination cursor"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page",
              "default": 20,
              "title": "Limit"
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChunkListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}/summary": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Summarize a document",
        "description": "Generate an LLM-powered summary of an indexed document.",
        "operationId": "get_document_summary",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Document ID",
              "title": "Document Id"
            },
            "description": "Document ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SummaryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}/similar": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Find similar documents",
        "description": "Find and analyze documents similar to the specified document.",
        "operationId": "get_similar_documents",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Document ID",
              "title": "Document Id"
            },
            "description": "Document ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimilarResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ChunkListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ChunkRecord"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "default": false
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "default": 20
          }
        },
        "type": "object",
        "title": "ChunkListResponse"
      },
      "ChunkRecord": {
        "properties": {
          "chunk_id": {
            "type": "string",
            "title": "Chunk Id",
            "default": ""
          },
          "text": {
            "type": "string",
            "title": "Text",
            "default": ""
          },
          "heading": {
            "type": "string",
            "title": "Heading",
            "default": ""
          },
          "chunk_index": {
            "type": "integer",
            "title": "Chunk Index",
            "default": 0
          }
        },
        "type": "object",
        "title": "ChunkRecord"
      },
      "DocumentDetailResponse": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id",
            "default": ""
          },
          "metadata": {
            "$ref": "#/components/schemas/DocumentMetadata"
          },
          "chunks": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Chunks"
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "default": ""
          }
        },
        "type": "object",
        "title": "DocumentDetailResponse"
      },
      "DocumentListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DocumentRecord"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "default": false
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "default": 20
          }
        },
        "type": "object",
        "title": "DocumentListResponse"
      },
      "DocumentMetadata": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "authors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Authors"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": ""
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          },
          "doi": {
            "type": "string",
            "title": "Doi",
            "default": ""
          }
        },
        "type": "object",
        "title": "DocumentMetadata"
      },
      "DocumentRecord": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id",
            "default": ""
          },
          "metadata": {
            "$ref": "#/components/schemas/DocumentMetadata"
          },
          "chunk_count": {
            "type": "integer",
            "title": "Chunk Count",
            "default": 0
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "default": ""
          }
        },
        "type": "object",
        "title": "DocumentRecord"
      },
      "DownloadRequest": {
        "properties": {
          "query": {
            "type": "string",
            "minLength": 1,
            "title": "Query",
            "description": "Search and download query"
          },
          "max_results": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Max Results",
            "description": "Maximum articles to download",
            "default": 5
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "DownloadRequest"
      },
      "DownloadResponse": {
        "properties": {
          "articles": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Articles"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "default": 0
          },
          "elapsed_ms": {
            "type": "number",
            "title": "Elapsed Ms",
            "default": 0.0
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "title": "DownloadResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IngestDocumentResponse": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id",
            "default": ""
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": ""
          },
          "chunks": {
            "type": "integer",
            "title": "Chunks",
            "default": 0
          }
        },
        "type": "object",
        "title": "IngestDocumentResponse"
      },
      "IngestRequest": {
        "properties": {
          "query": {
            "type": "string",
            "minLength": 1,
            "title": "Query",
            "description": "Search and ingest query"
          },
          "max_results": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Max Results",
            "description": "Maximum articles to ingest",
            "default": 5
          },
          "download_dir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Download Dir",
            "description": "Custom download directory"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "IngestRequest"
      },
      "IngestResponse": {
        "properties": {
          "documents": {
            "items": {
              "$ref": "#/components/schemas/IngestDocumentResponse"
            },
            "type": "array",
            "title": "Documents"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "default": 0
          },
          "elapsed_ms": {
            "type": "number",
            "title": "Elapsed Ms",
            "default": 0.0
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "title": "IngestResponse"
      },
...
      "SimilarResponse": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id",
            "default": ""
          },
          "analysis": {
            "type": "string",
            "title": "Analysis",
            "default": ""
          },
          "related_documents": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Related Documents"
          },
          "llm_model": {
            "type": "string",
            "title": "Llm Model",
            "default": ""
          },
          "llm_provider": {
            "type": "string",
            "title": "Llm Provider",
            "default": ""
          },
          "elapsed_ms": {
            "type": "number",
            "title": "Elapsed Ms",
            "default": 0.0
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "title": "SimilarResponse"
      },
      "SummaryResponse": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id",
            "default": ""
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "default": ""
          },
          "llm_model": {
            "type": "string",
            "title": "Llm Model",
            "default": ""
          },
          "llm_provider": {
            "type": "string",
            "title": "Llm Provider",
            "default": ""
          },
          "elapsed_ms": {
            "type": "number",
            "title": "Elapsed Ms",
            "default": 0.0
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "title": "SummaryResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}